Added validate user session endpoint

+++ src/services/deleteUser.service.js
@@ -1,7 +1,7 @@
 const db = require('./db.service');
 const Error = require('./domain/buisnessErrror.domain');
 const Success = require('./domain/success.domain');
-const { status } = require('../configs/general.config');
+const userManager = require('./userManager.service');
 
 class UserDeletedSuccess extends Success {
     constructor() {
@@ -12,19 +12,14 @@ class UserDeletedSuccess extends Success {
 }
 
 async function deleteUser(user) {
-    const userExists = await db.query(`SELECT * FROM Users WHERE username = ? AND status = ?`, [user.username, status.active]);
-    switch (true) {
-        case userExists.result.length > 0:
-            const deleteUserResult = await db.query(`DELETE FROM Users WHERE username = ?`, [user.username]);
-            switch (true) {
-                case deleteUserResult.result.affectedRows > 0:
-                    return new UserDeletedSuccess();
-                default:
-                    return new Error.DeleteUserError(); // Error occured while deleting user
-            }
-        default:
-            return new Error.UserDoesNotExist(); // User does not exist
-    }
+    const validateUserExistsResult = await userManager.validateUserExists(user.username);
+    if (validateUserExistsResult instanceof Error.BusinessError) return validateUserExistsResult;
+
+    const deleteUserResult = await db.query(`DELETE FROM Users WHERE username = ?`, [user.username]);
+    if (deleteUserResult instanceof Error.BusinessError) return deleteUserResult;
+    if (deleteUserResult.result.affectedRows > 0) return new UserDeletedSuccess(); // User deleted successfully
+
+    return new Error.DeleteUserError(); // Error occured while deleting user
 }
 
 module.exports = {
